refactor: folders/files structure
