+++ client/src/app/[lang]/page.tsx
@@ -1,7 +1,7 @@
 import { getDictionary } from '@/shared/localization/dictionaries';
 import { Locales } from '@/shared/types';
 
-import { TodoList } from '../../components/TodoList';
+import { TodoList } from '../../components/TodoList/TodoList';
 
 export default async function Home({ params }: { params: { lang: Locales } }) {
   const dictionary = await getDictionary(params.lang);
