+++ client/src/components/Todo/index.ts
@@ -1,0 +1,1 @@
+export { Todo } from './Todo';
