+++ client/src/components/TodoForm/index.ts
@@ -1,0 +1,1 @@
+export { TodoForm } from './TodoForm';
