+++ client/src/components/Todos/Todos.tsx
@@ -3,9 +3,9 @@
 import React, { useState, FC } from 'react';
 import axios from 'axios';
 
-import { TodoForm } from './TodoForm';
 import { Todo as TodoInterface } from '@/shared/types';
-import { Todo } from './Todo';
+import { TodoForm } from '../TodoForm';
+import { Todo } from '../Todo';
 
 interface TodoProps {
   todos: TodoInterface[];
