+++ client/src/components/Todos/index.ts
@@ -1,0 +1,1 @@
+export { Todos } from './Todos';
